Updated gray timer and end day msg
Gray timer editable text no longer has the underline and the pink flashing line is invisible by default and re appears once the text is clicked and turns invisible again after invisible when the user presses done.

The end day message now displays the percentage of time for every timer that was used during the day.