Added save preferences
save the past seven days and what the current day is(1-7). 
The key is the date and time when the user clicked end day.
The value is a string containing the seconds of each timer separated by a comma.("#,#,#,#,#,#,#,#")
Also calculated the percentage of each category and displayed it inside a textView.